kafka/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.34
botocore==1.31.34
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
multidict==6.0.4
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
xmltodict==0.13.0
yarl==1.9.2
slack_sdk==3.21.3